% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_config.R
\name{get_data_root}
\alias{get_data_root}
\title{Resolve quantamental data root}
\usage{
get_data_root()
}
\value{
character path
}
\description{
Priority:
\enumerate{
\item DATA_ROOT environment variable
\item Package snapshot in inst/extdata
}
}
