% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_arrow.R
\name{read_asset_parquet}
\alias{read_asset_parquet}
\title{Read asset-level parquet data}
\usage{
read_asset_parquet(asset_input, years = NULL, cols = NULL)
}
\arguments{
\item{years}{integer vector or NULL}

\item{cols}{character vector or NULL}

\item{asset_id}{character}
}
\value{
tibble
}
\description{
Read asset-level parquet data
}
