% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_semantics.R
\name{read_quantamentals}
\alias{read_quantamentals}
\title{Read quantamentals}
\usage{
read_quantamentals(asset_input)
}
\arguments{
\item{asset_input}{Ticker}
}
\value{
A tibble with flat, tabular quantamentals variables
}
\description{
Read quantamentals
}
