% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal_metadata.R
\name{read_asset_fundamentals}
\alias{read_asset_fundamentals}
\title{Read asset fundamentals}
\usage{
read_asset_fundamentals(asset_input)
}
\arguments{
\item{asset_input}{Ticker}
}
\value{
A tibble with flat, tabular fundamentals
}
\description{
Read asset fundamentals
}
